Major fix for mixed encoding filesystems
Sometimes it's possible (on unix) to end up with a filesystem where the
metadata is encoded with UTF8 but also with other encodings (such as
Latin1). This causes problems when Python is told that everything is
UTF8 and the endresult is that it would crash.

The change cannot resolve the underlying issue, BUT, it will warn the
user and avoid the file until fixed. To fix, just rename it and the
metadata is updated to UTF8 and all is well.

This has also been added to the readme so users won't get lost.