Fixtravis (#7)
travis filesystem produced a different order of files,
triggering a known issue in iceshelf. The issue is:

1. Move file A to B
2. Copy B to C
3. Run backup

iceself will detect that either B or C is identical to previous file
A, but since it doesn't do deduplication yet, it will select ONE at
random and then add the other one as a new one.

This gets the desired result, but the testing is more complicated.
In the future, deduplication will handle this case. As for end-users,
it will not cause any issue.